Move .sass and .scss into their own files
